Refactoring Data Types and removing all warnings from build.

+++ Breakout_32x32/Breakout_32x32/General.cpp
@@ -4,6 +4,6 @@
 
 #include "General.h"
 
-unsigned long g_Tic;
+uint32 g_Tic;
